Changed linear algebra to work using Armadillo. Added unit testing via GTest. Moved complext into the Gadgetron namespace.
Former-commit-id: 1137f4f9723f5ab712c0e6245482527b30ad5425

+++ gadgets/core/CropAndCombineGadget.cpp
@@ -48,7 +48,7 @@ process( GadgetContainerMessage<ISMRMRD::ImageHeader>* m1,
 	  mag += mag_tmp;
 	}
 
-	d2[offset_2] = std::polar((float)sqrt(mag),phase);
+	d2[offset_2] = std::polar(std::sqrt(mag),phase);
       }
     }
   }
